+++ src/TwinCatGateway.Contracts/OperationKind.cs
@@ -5,6 +5,7 @@ public enum OperationKind
     OpenSolution,
     Build,
     Synchronize,
+    CloseXae,
     Activate,
     RecoverToConfig,
     Test,